change: 로또 구입 가격의 범위 조정
- long 타입은 너무 큰 범위를 다루게 되어 관리가 어려워짐
- 따라서 로또 구입 가격의 범위는 int로 제한

+++ src/main/java/lotto/Application.java
@@ -10,8 +10,8 @@ public class Application {
         getBonusNumber();
     } 
     
-    private static long getPurchaseCost() {
-        return (long)getParsedUserInput(UserInputType.PURCHASE_COST);
+    private static int getPurchaseCost() {
+        return (int)getParsedUserInput(UserInputType.PURCHASE_COST);
     }
     
     private static List<Integer> getWinningNumbers() {
